backend/requirements.txt
pydicom
pillow
numpy
requests
fastapi
uvicorn
python-multipart
gdcm
pylibjpeg
pylibjpeg-libjpeg